[REFACTOR] Move dto related files to dto directory
